+++ examples/04-logger/src/app/modules/some/some.module.ts
@@ -1,11 +1,11 @@
-import { LoggerConfig, Module, Providers } from '@ditsmod/core';
+import { Module, Providers } from '@ditsmod/core';
 
 import { SomeController } from './some.controller';
 
 @Module({
   controllers: [SomeController],
   providersPerMod: [
-    ...new Providers().useValue(LoggerConfig, new LoggerConfig('trace'))
+    ...new Providers().useLogConfig({ level: 'trace' })
   ],
 })
 export class SomeModule {}