+++ examples/15-i18n/src/app/second/second.module.ts
@@ -1,19 +1,17 @@
-import { LoggerConfig, Module } from '@ditsmod/core';
-import { I18nModule, getI18nProviders, I18N_TRANSLATIONS } from '@ditsmod/i18n';
+import { Module } from '@ditsmod/core';
+import { I18nModule, I18nProviders, I18N_TRANSLATIONS } from '@ditsmod/i18n';
 
 import { FirstModule } from '../first/first.module';
 import { SecondController } from './second.controller';
 import { current } from './locales/current';
 import { imported } from './locales/imported';
 
-const loggerConfig = new LoggerConfig('debug');
-
 @Module({
   imports: [I18nModule, FirstModule],
   controllers: [SecondController],
   providersPerMod: [
-    { provide: LoggerConfig, useValue: loggerConfig },
-    ...getI18nProviders(this, { current, imported, moduleName: 'SecondModule' }, { defaultLng: 'uk' })
+    ...new I18nProviders()
+      .i18n({ current, imported }, { defaultLng: 'uk' }),
   ],
   exports: [I18N_TRANSLATIONS],
 })