+++ examples/15-i18n/src/app/third/third.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@ditsmod/core';
+import { DictService } from '@ditsmod/i18n';
+
+import { FirstModule } from '../first/first.module';
+import { MyDictService } from './dict.service';
+import { ThirdController } from './third.controller';
+
+@Module({
+  imports: [FirstModule],
+  controllers: [ThirdController],
+  providersPerReq: [{ provide: DictService, useClass: MyDictService }],
+})
+export class ThirdModule {}