+++ packages/core/src/index.ts
@@ -18,7 +18,7 @@ export { Route } from './decorators/route';
 export { Logger, LoggerConfig, LogLevel } from './types/logger';
 export { ControllerErrorHandler } from './services/controller-error-handler';
 export { BodyParserConfig } from './models/body-parser-config';
-export { ModConfig } from './models/mod-config';
+export { ModuleExtract } from './models/module-extract';
 export { ConsoleLogger } from './services/console-logger';
 export { Router } from './types/router';
 export { Status, getStatusText, isSuccess, STATUS_CODE_INFO } from './utils/http-status-codes';