+++ packages/core/src/module-factory.spec.ts
@@ -6,7 +6,7 @@ import { Controller, ControllerMetadata } from './decorators/controller';
 import { Module } from './decorators/module';
 import { RootModule } from './decorators/root-module';
 import { Route, RouteMetadata } from './decorators/route';
-import { ModConfig } from './models/mod-config';
+import { ModuleExtract } from './models/module-extract';
 import { NormalizedModuleMetadata } from './models/normalized-module-metadata';
 import { ModuleFactory } from './module-factory';
 import { defaultProvidersPerApp } from './services/default-providers-per-app';
@@ -74,7 +74,7 @@ describe('ModuleFactory', () => {
     ]);
     mock = injectorPerApp.get(MockModuleFactory);
     const logManager = new LogManager();
-    moduleManager = new ModuleManager(new LogMediator(logManager));
+    moduleManager = new ModuleManager(new LogMediator(logManager, {moduleName: 'fakeName'}));
   });
 
   describe('exportGlobalProviders()', () => {
@@ -322,13 +322,13 @@ describe('ModuleFactory', () => {
         mock.bootstrap([], new GlobalProviders(), '', Module3, moduleManager, new Set());
 
         const mod0 = mock.appMetadataMap.get(Module0);
-        const providerPerMod0: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+        const providerPerMod0: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module0' } };
         expect(mod0?.meta.providersPerMod).toEqual([providerPerMod0, Provider0]);
         expect(mod0?.meta.providersPerReq).toEqual([]);
         expect(mod0?.meta.ngMetadataName).toBe('Module');
 
         const mod1 = mock.appMetadataMap.get(Module1);
-        const providerPerMod1: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+        const providerPerMod1: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module1' } };
         expect(mod1?.meta.providersPerMod).toEqual([providerPerMod1, Provider1, Provider2, Provider3]);
 
         const tokensPerMod = getImportedTokens(mod1?.importedTokensMap.perMod);
@@ -338,7 +338,7 @@ describe('ModuleFactory', () => {
         expect(mod1?.meta.ngMetadataName).toBe('Module');
 
         const mod2 = mock.appMetadataMap.get(Module2);
-        const providerPerMod2: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+        const providerPerMod2: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module2' } };
         expect(mod2?.meta.providersPerMod).toEqual([providerPerMod2, Provider4, Provider5, Provider6]);
 
         const tokensPerMod2 = getImportedTokens(mod2?.importedTokensMap.perMod);
@@ -348,7 +348,7 @@ describe('ModuleFactory', () => {
         expect(mod2?.meta.ngMetadataName).toBe('Module');
 
         const mod3 = mock.appMetadataMap.get(Module3);
-        const providerPerMod3: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+        const providerPerMod3: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module3' } };
         expect(mod3?.meta.providersPerMod).toEqual([providerPerMod3]);
 
         // expect(mod3.providersPerReq).toEqual([Ctrl, [], Provider8, Provider9, overriddenProvider8]);
@@ -375,7 +375,7 @@ describe('ModuleFactory', () => {
 
         expect(mock.prefixPerMod).toBe('other');
         // expect(mock.router.find('GET', '/some/other').handle().controller).toBe(Ctrl);
-        const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: 'other' } };
+        const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: 'other', moduleName: 'Module4' } };
         expect(mock.meta.providersPerMod).toEqual([providerPerMod]);
 
         expect(mock?.importedProvidersPerMod).toBeDefined();