+++ packages/core/src/services/app-initializer.spec.ts
@@ -13,7 +13,7 @@ import { LogFilter, LogMediator } from './log-mediator';
 import { ModuleManager } from './module-manager';
 import { Extension, ModuleType, ModuleWithParams, ServiceProvider } from '../types/mix';
 import { Controller } from '../decorators/controller';
-import { ModConfig } from '../models/mod-config';
+import { ModuleExtract } from '../models/module-extract';
 import { ImportObj, MetadataPerMod1 } from '../types/metadata-per-mod';
 import { RootMetadata } from '../models/root-metadata';
 
@@ -63,7 +63,7 @@ describe('AppInitializer', () => {
 
   describe('prepareProvidersPerApp()', () => {
     beforeEach(() => {
-      const logMediator = new LogMediator(new LogManager());
+      const logMediator = new LogMediator(new LogManager(), {moduleName: 'fakeName'});
       moduleManager = new ModuleManager(logMediator);
       const rootMeta = new RootMetadata();
       mock = new AppInitializerMock(rootMeta, moduleManager, logMediator);
@@ -297,7 +297,7 @@ describe('AppInitializer', () => {
     class AppModule {}
 
     beforeEach(() => {
-      const logMediator = new LogMediator(new LogManager());
+      const logMediator = new LogMediator(new LogManager(), {moduleName: 'fakeName'});
       moduleManager = new ModuleManager(logMediator);
       const rootMeta = new RootMetadata();
       mock = new AppInitializerMock(rootMeta, moduleManager, logMediator);
@@ -415,7 +415,7 @@ describe('AppInitializer', () => {
     let appMetadataMap: Map<AnyModule, MetadataPerMod1>;
 
     beforeAll(() => {
-      const logMediator = new LogMediator(new LogManager());
+      const logMediator = new LogMediator(new LogManager(), {moduleName: 'fakeName'});
       moduleManager = new ModuleManager(logMediator);
       const rootMeta = new RootMetadata();
       mock = new AppInitializerMock(rootMeta, moduleManager, logMediator);
@@ -458,7 +458,7 @@ describe('AppInitializer', () => {
     it('Module0', async () => {
       const mod0 = appMetadataMap.get(Module0);
       expect(mod0?.meta.providersPerApp).toEqual([]);
-      const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+      const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module0' } };
       expect(mod0?.meta.providersPerMod).toEqual([providerPerMod, Provider0]);
       expect(mod0?.meta.providersPerReq).toEqual([]);
       checkGlobalProviders(mod0);
@@ -467,7 +467,7 @@ describe('AppInitializer', () => {
     it('Module1', async () => {
       const mod1 = appMetadataMap.get(Module1);
       expect(mod1?.meta.providersPerApp).toEqual([]);
-      const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+      const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module1' } };
       expect(mod1?.meta.providersPerMod).toEqual([providerPerMod, obj1, Provider2]);
       checkGlobalProviders(mod1);
     });
@@ -475,7 +475,7 @@ describe('AppInitializer', () => {
     it('Module2', async () => {
       const mod2 = appMetadataMap.get(module2WithParams);
       expect(mod2?.meta.providersPerApp).toEqual([]);
-      const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+      const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module2' } };
       expect(mod2?.meta.providersPerMod).toEqual([providerPerMod, Provider3, Provider4]);
       expect(mod2?.meta.providersPerReq).toEqual([]);
       checkGlobalProviders(mod2);
@@ -484,7 +484,7 @@ describe('AppInitializer', () => {
     it('Module3', async () => {
       const mod3 = appMetadataMap.get(module3WithParams);
       expect(mod3?.meta.providersPerApp).toEqual([]);
-      const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: 'one' } };
+      const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: 'one', moduleName: 'Module3' } };
       expect(mod3?.meta.providersPerMod).toEqual([providerPerMod]);
       expect(mod3?.meta.providersPerReq).toEqual([Provider5, Provider6, Provider7]);
       checkGlobalProviders(mod3);
@@ -495,7 +495,7 @@ describe('AppInitializer', () => {
       const appMetadataMap = mock.bootstrapModuleFactory(moduleManager);
       const mod4 = appMetadataMap.get(module4WithParams);
       expect(mod4?.meta.providersPerApp).toEqual([]);
-      const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+      const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'Module4' } };
       expect(mod4?.meta.providersPerMod).toEqual([providerPerMod]);
       expect(mod4?.meta.providersPerReq).toEqual([Provider8, Provider9]);
       checkGlobalProviders(mod4);
@@ -506,7 +506,7 @@ describe('AppInitializer', () => {
       const appMetadataMap = mock.bootstrapModuleFactory(moduleManager);
       const root1 = appMetadataMap.get(AppModule);
       expect(root1?.meta.providersPerApp.slice(0, 2)).toEqual([Logger, { provide: Router, useValue: 'fake' }]);
-      const providerPerMod: ServiceProvider = { provide: ModConfig, useValue: { prefixPerMod: '' } };
+      const providerPerMod: ServiceProvider = { provide: ModuleExtract, useValue: { path: '', moduleName: 'AppModule' } };
       expect(root1?.meta.providersPerMod).toEqual([providerPerMod]);
       expect(root1?.meta.providersPerReq).toEqual([]);
       checkGlobalProviders(root1);
@@ -534,7 +534,7 @@ describe('AppInitializer', () => {
       }
 
       // Simulation of a call from the Application
-      const logMediator = new LogMediatorMock(new LogManager());
+      const logMediator = new LogMediatorMock(new LogManager(), {moduleName: 'fakeName'});
       moduleManager = new ModuleManager(logMediator);
       const rootMeta = new RootMetadata();
       mock = new AppInitializerMock(rootMeta, moduleManager, logMediator);
@@ -585,7 +585,7 @@ describe('AppInitializer', () => {
 
     beforeEach(() => {
       testMethodSpy.mockRestore();
-      const logMediator = new LogMediator(new LogManager());
+      const logMediator = new LogMediator(new LogManager(), {moduleName: 'fakeName'});
       moduleManager = new ModuleManager(logMediator);
       const rootMeta = new RootMetadata();
       mock = new AppInitializerMock(rootMeta, moduleManager, logMediator);
@@ -628,7 +628,7 @@ describe('AppInitializer', () => {
 
     beforeEach(() => {
       jestFn.mockRestore();
-      const logMediator = new LogMediator(new LogManager());
+      const logMediator = new LogMediator(new LogManager(), {moduleName: 'fakeName'});
       moduleManager = new ModuleManager(logMediator);
       const rootMeta = new RootMetadata();
       mock = new AppInitializerMock(rootMeta, moduleManager, logMediator);