+++ packages/core/src/utils/providers.spec.ts
@@ -1,9 +1,10 @@
 import { describe, it, fit, expect, jest } from '@jest/globals';
-import { Logger } from '../types/logger';
+import { Logger, LoggerConfig } from '../types/logger';
 import { ServiceProvider } from '../types/mix';
 
 import { ConsoleLogger } from '../services/console-logger';
 import { Providers } from './providers';
+import { LogFilter, LogMediator } from '../services/log-mediator';
 
 describe('Providers', () => {
   it('call constuctor not to throw', () => {
@@ -12,7 +13,7 @@ describe('Providers', () => {
 
   it('works useAnyValue()', () => {
     const value = new Providers().useAnyValue('token', 'value');
-    expect([...value]).toEqual([{ provide: 'token', useValue: 'value', multi: undefined }]);
+    expect([...value]).toEqual([{ provide: 'token', useValue: 'value' }]);
   });
 
   it('works useValue()', () => {
@@ -20,7 +21,7 @@ describe('Providers', () => {
       one: string;
     }
     const value = new Providers().useValue(A, { one: 'value' });
-    expect([...value]).toEqual([{ provide: A, useValue: { one: 'value' }, multi: undefined }]);
+    expect([...value]).toEqual([{ provide: A, useValue: { one: 'value' } }]);
   });
 
   it('works useClass()', () => {
@@ -32,21 +33,41 @@ describe('Providers', () => {
       two: number;
     }
     const value = new Providers().useClass(A, B);
-    expect([...value]).toEqual([{ provide: A, useClass: B, multi: undefined }]);
+    expect([...value]).toEqual([{ provide: A, useClass: B }]);
   });
 
   it('works useLogger()', () => {
     const logger = new ConsoleLogger();
     const value = new Providers().useLogger(logger);
-    expect([...value]).toEqual([{ provide: Logger, useValue: logger, multi: undefined }]);
+    expect([...value]).toEqual([{ provide: Logger, useValue: logger }]);
+  });
+
+  it('works useLogConfig()', () => {
+    const loggerConfig = new LoggerConfig();
+
+    const config1 = new Providers().useLogConfig(loggerConfig);
+    expect([...config1]).toEqual([{ provide: LoggerConfig, useValue: loggerConfig }]);
+
+    const config2 = new Providers().useLogConfig(loggerConfig, { tags: ['one'] });
+    expect([...config2]).toEqual([
+      { provide: LoggerConfig, useValue: loggerConfig },
+      { provide: LogFilter, useValue: { tags: ['one'] } },
+    ]);
+  });
+
+  it('works useLogMediator()', () => {
+    class CustomLogMediator extends LogMediator {}
+
+    const config1 = new Providers().useLogMediator(CustomLogMediator);
+    expect([...config1]).toEqual([CustomLogMediator, { provide: LogMediator, useExisting: CustomLogMediator }]);
   });
 
   it('works multi calling', () => {
     const logger = new ConsoleLogger();
     const value = new Providers().useLogger(logger).useAnyValue('token', 'value');
     const expectedArr: ServiceProvider[] = [
-      { provide: Logger, useValue: logger, multi: undefined },
-      { provide: 'token', useValue: 'value', multi: undefined },
+      { provide: Logger, useValue: logger },
+      { provide: 'token', useValue: 'value' },
     ];
     expect([...value]).toEqual(expectedArr);
   });
@@ -78,15 +99,7 @@ describe('Providers', () => {
     const providers = new Providers();
 
     function callback() {
-      providers
-        .use(Some)
-        .use(Other)
-        .two()
-        .two()
-        .one('Mostia')
-        .use(Third)
-        .three()
-        .useAnyValue('token', 'value');
+      providers.use(Some).use(Other).two().two().one('Mostia').use(Third).three().useAnyValue('token', 'value');
     }
 
     expect(callback).not.toThrow();
@@ -94,8 +107,8 @@ describe('Providers', () => {
     expect(Other.prototype.two).toBeCalledTimes(2);
     expect(Third.prototype.three).toBeCalledTimes(1);
     expect([...providers]).toEqual([
-      { provide: 'Mostia', useValue: 'молоток', multi: undefined },
-      { provide: 'token', useValue: 'value', multi: undefined },
+      { provide: 'Mostia', useValue: 'молоток' },
+      { provide: 'token', useValue: 'value' },
     ]);
   });
 });