+++ packages/i18n/src/i18n.module.ts
@@ -1,4 +1,4 @@
-import { Module, PRE_ROUTER_EXTENSIONS } from '@ditsmod/core';
+import { Module, PRE_ROUTER_EXTENSIONS, Providers } from '@ditsmod/core';
 
 import { I18nLogMediator } from './i18n-log-mediator';
 import { I18nExtension } from './i18n.extension';
@@ -8,7 +8,7 @@ import { DictPerModService } from './dict-per-mod.service';
 
 @Module({
   extensions: [[I18N_EXTENSIONS, PRE_ROUTER_EXTENSIONS, I18nExtension, true]],
-  providersPerMod: [I18nOptions, I18nLogMediator, DictPerModService],
+  providersPerMod: [I18nOptions, DictPerModService, ...new Providers().useLogMediator(I18nLogMediator)],
   providersPerReq: [DictService],
   exports: [DictService],
 })