+++ packages/router/CHANGELOG.md
@@ -1,5 +1,12 @@
+<a name="router-2.1.0"></a>
+# [router-2.1.0](https://github.com/ts-stack/ditsmod/releases/tag/router-2.1.0) (2022-08-27)
+
+### BREAKING CHANGES
+
+- Migration to Ditsmod v2.11.0.
+
 <a name="router-2.0.0-next.0"></a>
-# [router-2.0.0-next.0](https://github.com/ts-stack/ditsmod/releases/tag/router-2.0.0-next.0) (2021-11-20)
+## [router-2.0.0-next.0](https://github.com/ts-stack/ditsmod/releases/tag/router-2.0.0-next.0) (2021-11-20)
 
 ### BREAKING CHANGES